requirements.txt
pdfplumber
chromadb
sentence-transformers
torch
transformers
python-dotenv
tqdm
typing-extensions
accelerate
